add if to check file existing

+++ aeme/scripts/run.sh
@@ -7,5 +7,7 @@ MODEL="ae AEModel"
 `export PATH=$PATH:/usr/local/cuda/bin`
 `export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/usr/local/cuda/lib:/usr/local/cuda/lib64`
 
-`rm nohup.out`
+if [ -f "nohup.out" ]; then
+    `rm nohup.out`
+fi
 `nohup python ../execute.py --module-path ~/aeme -m $MODEL -i $IN -o $OUT -b 128 -e 1000 --noise-type SP --noise-ratio 0.05 &`
